Refactor MediaCarousel into smaller components
The MediaCarousel component was refactored into smaller, more manageable components to improve code organization and maintainability.
