+++ src/components/media/EnlargedImageView.tsx
@@ -1,0 +1,80 @@
+
+import React from "react";
+import { Dialog, DialogContent } from "@/components/ui/dialog";
+import { Button } from "@/components/ui/button";
+import { Crop, Trash2, X } from "lucide-react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+
+interface EnlargedImageViewProps {
+  selectedMedia: string | null;
+  onClose: () => void;
+  onStartCrop: () => void;
+  onDelete: () => void;
+}
+
+const EnlargedImageView = ({
+  selectedMedia,
+  onClose,
+  onStartCrop,
+  onDelete,
+}: EnlargedImageViewProps) => {
+  if (!selectedMedia) return null;
+
+  return (
+    <Dialog open={!!selectedMedia} onOpenChange={onClose}>
+      <DialogContent className="max-w-[90vw] max-h-[90vh] p-0" hideCloseButton>
+        <div className="relative">
+          <img
+            src={selectedMedia}
+            alt="Enlarged view"
+            className="w-full h-full object-contain"
+          />
+          <div className="absolute top-2 right-2 flex gap-2">
+            <Button size="icon" variant="secondary" onClick={onStartCrop}>
+              <Crop className="h-4 w-4" />
+            </Button>
+            <AlertDialog>
+              <AlertDialogTrigger asChild>
+                <Button size="icon" variant="destructive">
+                  <Trash2 className="h-4 w-4" />
+                </Button>
+              </AlertDialogTrigger>
+              <AlertDialogContent>
+                <AlertDialogHeader>
+                  <AlertDialogTitle>Are you sure?</AlertDialogTitle>
+                  <AlertDialogDescription>
+                    This action cannot be undone. This will permanently delete this media.
+                  </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                  <AlertDialogCancel>Cancel</AlertDialogCancel>
+                  <AlertDialogAction
+                    onClick={onDelete}
+                    className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+                  >
+                    Delete
+                  </AlertDialogAction>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
+            <Button size="icon" variant="secondary" onClick={onClose}>
+              <X className="h-4 w-4" />
+            </Button>
+          </div>
+        </div>
+      </DialogContent>
+    </Dialog>
+  );
+};
+
+export default EnlargedImageView;
